Various readme edits
Add required OMSI version:
This tool did not work when I tried it on version 2.2.032 (to be expected, but I did not realise I was on an older version). 
Although most users are probably on the latest version, I've added a prerequisite of OMSI 2 version 2.3.004 to try and mitigate any confusion.

This program must be ran as administrator, so I have changed some wording.